Split Facility overview into 2 pages
